+++ plotnine/stats/stat_function.py
@@ -37,15 +37,15 @@ class stat_function(stat):
 
     ::
 
-        '..x..'  # x points are which the function is evaluated
-        '..y..'  # Points evaluated at x
+        'x'  # x points are which the function is evaluated
+        'y'  # Points evaluated at x
     """
     DEFAULT_PARAMS = {'geom': 'path', 'position': 'identity',
                       'na_rm': False,
                       'fun': None, 'n': 101, 'args': None,
                       'xlim': None}
 
-    DEFAULT_AES = {'y': '..y..'}
+    DEFAULT_AES = {'y': 'calc(y)'}
     CREATES = {'y'}
 
     @classmethod