+++ plotnine/stats/stat_sum.py
@@ -26,14 +26,14 @@ class stat_sum(stat):
 
     ::
 
-        '..n..'     # Number of observations at a position
-        '..prop..'  # Ratio of points in that panel at a position
+        'n'     # Number of observations at a position
+        'prop'  # Ratio of points in that panel at a position
 
     """
     REQUIRED_AES = {'x', 'y'}
     DEFAULT_PARAMS = {'geom': 'point', 'position': 'identity',
                       'na_rm': False}
-    DEFAULT_AES = {'size': '..prop..', 'weight': 1}
+    DEFAULT_AES = {'size': 'calc(prop)', 'weight': 1}
     CREATES = {'n', 'prop'}
 
     @classmethod